+ bfloat16

+++ benchmarks/triton_attention/v1_offline_flash_attn.py
@@ -92,6 +92,6 @@ if __name__ == "__main__":
             f = executor.submit(benchmark_vllm, args)
             f.result()
 
-    for dtype in ["float16", "float32"]:
+    for dtype in ["float16", "bfloat16"]:
         args.dtype = dtype
         run(args)
